+++ src/main/java/lermitage/intellij/extra/icons/ExtraIconProvider.java
@@ -90,7 +90,8 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                 ofDir("dir_circleci", "extra-icons/folder_circleci.svg", "Circle CI: .circleci folder")
                     .eq(".circleci"),
                 ofDir("dir_dependabot", "extra-icons/folder_dependabot.svg", "Dependabot: .dependabot folder")
-                    .eq(".dependabot"),
+                    .eq(".dependabot")
+                    .altIcons("extra-icons/folder_dependabot_alt.svg"),
                 ofDir("dir_docker", "extra-icons/folder_docker.svg", "Docker: docker folder")
                     .eq("docker")
                     .tags(ModelTag.DOCKER),
@@ -418,9 +419,11 @@ public class ExtraIconProvider extends BaseIconProvider implements DumbAware {
                 // file plus its containing folder
                 //
                 ofFile("dependabot_config", "extra-icons/dependabot.svg", "Dependabot: .dependabot/config.yml")
-                    .eq("config.yml").parents(".dependabot"),
+                    .eq("config.yml").parents(".dependabot")
+                    .altIcons("extra-icons/dependabot_alt.svg"),
                 ofFile("dependabot", "extra-icons/dependabot.svg", "Dependabot: .github/dependabot.yml")
-                    .eq("dependabot").end(YML).parents(".github"),
+                    .eq("dependabot").end(YML).parents(".github")
+                    .altIcons("extra-icons/dependabot_alt.svg"),
                 ofFile("golwroot", "extra-icons/glowroot.png", "Glowroot: glowroot/admin.json and glowroot/config.json")
                     .eq("admin.json", "config.json").parents("glowroot"),
                 ofFile("jgitver", "extra-icons/jgitver.svg", "jgitver: .mvn/jgitver.config.xml")