+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2022.1.10 (WIP)
 * add an option in Extra Icons settings in order to force icons reloading on demand. Use it if you still see errors when querying IDE filename index: wait until indexing is done, go to File, Settings, Appearance & Behavior, Extra Icons, then hit the `Reload projects icons` button. Again, feel free to upvote [**IDEA-289822**](https://youtrack.jetbrains.com/issue/IDEA-289822), it would help.
+* add an alternative Dependabot icon.
 
 ## 2022.1.9 (2022/08/27)
 * support many audio/video files. Thx to contributor **Serial-ATA**.